Wrap the Cargo linker to not invalidate Cargo cache with changing SONAME.
This avoids
https://github.com/mozilla/application-services/issues/382.  Suppose
your Cargo workspace has two crates which share some dependencies.
The existing invocation, for each crate, includes a `-C link-arg=...`
that is different for each crate.  Even though it's link-only, this
forces the recompilation of each dependency for each workspace crate.
This is a complicated way of hiding the choice of precise link
argument from Cargo, so that it's only incorporated in the final
linking step and not into the dependency graph proper.