Guard against misconfigured target directory
When the Rust project is inside a Cargo Workspace this plugin will
_fail_ silently because by default it assumes Cargo target directory
will be inside the module directory.

This patch add a guard against misconfigured target directory. When the
target directory not exists, we'll make the build fails and point user
to the documentation.

```
Cargo target directory (`../../target`) does not exists. If you're using
a Cargo Workspace, you need to specify the target directory in
`build.gradle` or `local.properties` (see
https://github.com/mozilla/rust-android-gradle#targetdirectory).
```